requirements.txt
autopep8==1.5.6
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.15.8
pycodestyle==2.7.0
pymongo==3.11.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
shiboken2==5.15.2
six==1.15.0
toml==0.10.2
